added breadth and depth to cascading sync

+++ Core/Types/Money.cs
@@ -18,7 +18,7 @@ namespace Trustcoin.Core.Types
         public static explicit operator Money(float value) => new Money(value);
         public static implicit operator float(Money money) => money.Value;
         public static Money operator +(Money m1, Money m2) => (Money)(m1.Value + m2.Value);
-        public static Money operator *(int n, Money m) => (Money)(n * m.Value);
+        public static Money operator *(float n, Money m) => (Money)(n * m.Value);
         public static Money operator -(Money m1, Money m2) => (Money)(m1.Value - m2.Value);
 
         public override string ToString() => $"${Value}";
